add veci_test.rb #6
